# mdw_backend_todo/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Documentation",
        "description": "Documentation for your API",
        "contact": {
            "name": "Your Name",
            "email": "your.email@example.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/task": {
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Create a new task",
                "description": "Create a new task with the provided details.",
                "operationId": "createTask",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "due_date",
                                    "remind_at"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Task title",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Task description",
                                        "type": "string"
                                    },
                                    "due_date": {
                                        "description": "Due date of the task (YYYY-MM-DD HH:mm:ss)",
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "remind_at": {
                                        "description": "Remind date of the task (YYYY-MM-DD HH:mm:ss)",
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "status": {
                                        "description": "Task status, default is 'en attente'",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Task created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "title": [
                                                    "The title field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error creating task"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/task-statistics": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Get statistics for user tasks",
                "description": "Returns various statistics for tasks belonging to the authenticated user.",
                "operationId": "getTaskStatistics",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statistics": {
                                            "properties": {
                                                "en_attente": {
                                                    "description": "Count of tasks with status 'en attente'",
                                                    "type": "integer"
                                                },
                                                "open": {
                                                    "description": "Count of tasks with status 'open'",
                                                    "type": "integer"
                                                },
                                                "in_progress": {
                                                    "description": "Count of tasks with status 'in progress'",
                                                    "type": "integer"
                                                },
                                                "accepted": {
                                                    "description": "Count of tasks with status 'Accepted'",
                                                    "type": "integer"
                                                },
                                                "solved": {
                                                    "description": "Count of tasks with status 'solved'",
                                                    "type": "integer"
                                                },
                                                "on_hold": {
                                                    "description": "Count of tasks with status 'on hold'",
                                                    "type": "integer"
                                                },
                                                "overdue": {
                                                    "description": "Count of tasks that are overdue",
                                                    "type": "integer"
                                                },
                                                "to_do": {
                                                    "description": "Alias for 'en_attente'",
                                                    "type": "integer"
                                                },
                                                "open_tasks": {
                                                    "description": "Alias for 'open'",
                                                    "type": "integer"
                                                },
                                                "due_today": {
                                                    "description": "Count of tasks due today",
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    }
}